requirements.txt
groq==0.3.0
boto3==1.34.131
werkzeug==3.0.1
Flask==2.3.0
Flask-CORS==4.0.0
Flask-Mail==0.9.1
python-dotenv==1.0.0

